refactor(game): split trauma_producers, phobia/scan, anatomy, effects
Four files split into submodules in game/src/tributes/afflictions/:

trauma_producers.rs (778) -> producers/mod.rs + shared.rs + 4 producer files + tests.rs
phobia/scan.rs (761) -> scan.rs + outcomes.rs
anatomy.rs (686) -> anatomy.rs (122) + anatomy_tests.rs (564)
effects.rs (641) -> effects/mod.rs + stat_modifiers.rs + brain_bias.rs + trauma_effects.rs

Also: fixed flashback_chance duplicate in trauma.rs (now imports from effects)
