+++ game/src/tributes/afflictions/producers/survive_near_death.rs
@@ -1,0 +1,54 @@
+//! Producer (c): Survive near-death — Moderate trauma.
+
+use crate::games::Game;
+use crate::messages::{MessagePayload, Phase};
+use shared::afflictions::{DeathCause, Severity, TraumaSource};
+
+use super::shared::{MAX_HEALTH, TraumaEvent, apply_trauma_events};
+
+pub(super) fn produce_survive_near_death(game: &mut Game, phase: Phase) {
+    // Collect wound events that pushed a tribute to <= 10% HP
+    let mut events: Vec<TraumaEvent> = Vec::new();
+
+    for msg in &game.messages {
+        if msg.phase != phase {
+            continue;
+        }
+        let MessagePayload::TributeWounded {
+            victim,
+            attacker,
+            hp_lost: _,
+        } = &msg.payload
+        else {
+            continue;
+        };
+
+        let Some(tribute) = game
+            .tributes
+            .iter()
+            .find(|t| t.identifier == victim.identifier)
+        else {
+            continue;
+        };
+
+        let hp_after = tribute.attributes.health;
+        let hp_percent = hp_after * 100 / MAX_HEALTH;
+
+        if hp_percent <= 10 && tribute.is_alive() {
+            let death_cause = attacker
+                .as_ref()
+                .map(|a| DeathCause::Tribute(a.identifier.clone()))
+                .unwrap_or(DeathCause::Unknown);
+
+            events.push(TraumaEvent {
+                tribute_id: tribute.identifier.clone(),
+                tribute_name: tribute.name.clone(),
+                source: TraumaSource::NearDeath { cause: death_cause },
+                severity: Severity::Moderate,
+                cause_hint: String::new(),
+            });
+        }
+    }
+
+    apply_trauma_events(game, events, false);
+}
